+++ pc/std/tracker/Samples.h
@@ -28,7 +28,7 @@ struct Sample
   Brr *brr; // raw brr sample data. Have to iterate it to END block
   /* offset into sample where loop starts, specified by number of BRR
    * blocks or [by samples and must be divisible by 16]*/
-  size_t brrsize;
+  uint16_t brrsize;
   uint16_t rel_loop;
   /* signed offset in semitones to pitch the sample. This will be used
    * directly by the SNES driver. The tracker must impose the restraints
